requirements.txt
Django==4.2  # or your specific version
djangorestframework==3.14  # or your specific version
djangorestframework-simplejwt==5.2  # or your specific version
opencv-python==4.5.5.64  # or your specific version
asgiref==3.5.2
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2
opencv-python==4.5.5.64
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
# Add other dependencies as needed
